[chore] #41 change type (float -> int)

+++ src/main/java/com/wonnabe/product/domain/InsuranceProductVO.java
@@ -24,11 +24,11 @@ public class InsuranceProductVO {
     private BigDecimal malePremium;
 
     // Recommend를 위한 Score + 상세보기 Chart data
-    private Float scorePriceCompetitiveness; // 가격 경쟁력 점수
-    private Float scoreCoverageLimit; // 보장 한도 점수
-    private Float scoreCoverageScope; // 보장 범위 점수
-    private Float scoreDeductibleLevel; // 자기 부담금 점수
-    private Float scoreRefundScope; // 환급 범위 점수
+    private int scorePriceCompetitiveness; // 가격 경쟁력 점수
+    private int scoreCoverageLimit; // 보장 한도 점수
+    private int scoreCoverageScope; // 보장 범위 점수
+    private int scoreDeductibleLevel; // 자기 부담금 점수
+    private int scoreRefundScope; // 환급 범위 점수
 
     // --- 보험 상세보기를 위한 변수 ---
     private String coverageType;
